requirements.txt
flask
line-bot-sdk
numpy
google-generativeai
google-cloud-aiplatform
